Small modifications in the ODE system files
Right now, the issue for some of the results in Rossler seems to be the integration step delta_t. When delta_t is small enough, everything works properly; but "small enough" seems to be dependent on the target ODE system. delta_t=0.1 is enough for Lotka-Volterra, but not for Rossler, where good results are obtained with delta_t=0.01