Completed script to test different approximations with PySR
There are A LOT of possible approximations for the first-order derivative y'(t) of a function y(t). This script is testing the two main approximations, one computing delta_y/delta_t numerically, the other using the forward Euler's method to obtain numerical values of F_y. Each technique has some extra parameter, for example the order of the approximation, or the smoothing.